Louder! (feat. xrefresh and notify-send)
I've been using my watch timer at my dayjob lately, but it's loud—no
one has complained, but maybe they're just shy. Maybe this will
suffice as a "visual bell" that I can trust to relieve me when it's
time. (Distrusting the timer and feeling the need to pick at the time
left is contrary to the method!)